make it possible to use the ICE with bare metal simulators
Hopefully this makes sense and I didn't break picosim...

Also cleaned up #include's in z80core

Did remove the #ifdef PICO stuff from simcore. Why is this necessary?
I have the Pico SDK here and can build picosim without it.
