+++ picosim/CMakeLists.txt
@@ -13,6 +13,9 @@ include_directories(
 add_executable(picosim
 	picosim.c
 	../z80core/simcore.c
+	../z80core/simdis.c
+	../z80core/simglb.c
+	../z80core/simice.c
 	../z80core/simz80.c
 	../z80core/simz80-cb.c
 	../z80core/simz80-dd.c
@@ -20,7 +23,6 @@ add_executable(picosim
 	../z80core/simz80-ed.c
 	../z80core/simz80-fd.c
 	../z80core/simz80-fdcb.c
-	../z80core/simglb.c
 	config.c
 	memory.c
 	iosim.c
